+++ bigtree/tree/construct.py
@@ -24,12 +24,16 @@ from bigtree.utils.exceptions import (
 try:
     import pandas as pd
 except ImportError:  # pragma: no cover
-    pd = None
+    from unittest.mock import MagicMock
+
+    pd = MagicMock()
 
 try:
     import polars as pl
 except ImportError:  # pragma: no cover
-    pl = None
+    from unittest.mock import MagicMock
+
+    pl = MagicMock()
 
 __all__ = [
     "add_path_to_tree",