+++ bigtree/tree/export.py
@@ -28,22 +28,30 @@ from bigtree.utils.iterators import levelordergroup_iter, preorder_iter
 try:
     import pandas as pd
 except ImportError:  # pragma: no cover
-    pd = None
+    from unittest.mock import MagicMock
+
+    pd = MagicMock()
 
 try:
     import polars as pl
 except ImportError:  # pragma: no cover
-    pl = None
+    from unittest.mock import MagicMock
+
+    pl = MagicMock()
 
 try:
     import pydot
 except ImportError:  # pragma: no cover
-    pydot = None
+    from unittest.mock import MagicMock
+
+    pydot = MagicMock()
 
 try:
     from PIL import Image, ImageDraw, ImageFont
 except ImportError:  # pragma: no cover
-    Image = ImageDraw = ImageFont = None
+    from unittest.mock import MagicMock
+
+    Image = ImageDraw = ImageFont = MagicMock()
 
 
 __all__ = [
@@ -73,9 +81,10 @@ def print_tree(
     attr_omit_null: bool = False,
     attr_bracket: List[str] = ["[", "]"],
     style: Union[str, Iterable[str], BasePrintStyle] = "const",
-    **print_kwargs: Any,
+    **kwargs: Any,
 ) -> None:
     """Print tree to console, starting from `tree`.
+    Accepts kwargs for print() function.
 
     - Able to select which node to print from, resulting in a subtree, using `node_name_or_path`
     - Able to customize for maximum depth to print, using `max_depth`
@@ -90,8 +99,6 @@ def print_tree(
     - (List[str]): Choose own style for stem, branch, and final stem icons, they must have the same number of characters
     - (BasePrintStyle): `ANSIPrintStyle`, `ASCIIPrintStyle`, `ConstPrintStyle`, `ConstBoldPrintStyle`, `RoundedPrintStyle`,
     `DoublePrintStyle` style or inherit from `BasePrintStyle`
-
-    Remaining kwargs are passed without modification to python's `print` function.
 
     Examples:
         **Printing tree**
@@ -249,7 +256,7 @@ def print_tree(
             if attr_str:
                 attr_str = f" {attr_bracket_open}{attr_str}{attr_bracket_close}"
         node_str = f"{_node.node_name}{attr_str}"
-        print(f"{pre_str}{fill_str}{node_str}", **print_kwargs)
+        print(f"{pre_str}{fill_str}{node_str}", **kwargs)
 
 
 def yield_tree(
@@ -433,9 +440,10 @@ def hprint_tree(
     max_depth: int = 0,
     intermediate_node_name: bool = True,
     style: Union[str, Iterable[str], BaseHPrintStyle] = "const",
-    **print_kwargs: Any,
+    **kwargs: Any,
 ) -> None:
     """Print tree in horizontal orientation to console, starting from `tree`.
+    Accepts kwargs for print() function.
 
     - Able to select which node to print from, resulting in a subtree, using `node_name_or_path`
     - Able to customize for maximum depth to print, using `max_depth`
@@ -447,8 +455,6 @@ def hprint_tree(
     - (List[str]): Choose own style icons, they must have the same number of characters
     - (BaseHPrintStyle): `ANSIHPrintStyle`, `ASCIIHPrintStyle`, `ConstHPrintStyle`, `ConstBoldHPrintStyle`,
     `RoundedHPrintStyle`, `DoubleHPrintStyle` style or inherit from BaseHPrintStyle
-
-    Remaining kwargs are passed without modification to python's `print` function.
 
     Examples:
         **Printing tree**
@@ -549,7 +555,7 @@ def hprint_tree(
         max_depth=max_depth,
         style=style,
     )
-    print("\n".join(result), **print_kwargs)
+    print("\n".join(result), **kwargs)
 
 
 def hyield_tree(